refactoring routes for users and index routes
