+++ typeserver/routes/user/index.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const controller = require('./user.controller');
+
+const router = express.Router();
+
+router.get('/', controller.index);
+router.post('/', controller.addUser);
+
+module.exports = router;
